refactor(weaver-core): move GithubStatus PR-snapshot model into core
The per-branch pull-request snapshot is a plain serializable DB row served
under BranchView; it belongs with the other shared models, not in the loom
orchestrator. The gh-CLI polling that produces it stays in loom; only the
struct and its signature/event_data helpers move, so the wire DTOs in
weaver-api can reference one definition.
